refactor the compare files function
